feat: remove donation in process page|
