+++ src/config/routes/index.tsx
@@ -5,7 +5,6 @@ import CausesPage from "pages/donations/CausesPage";
 import CurrentUserProvider from "contexts/currentUserContext";
 import ConfirmEmailPage from "pages/donations/ConfirmEmailPage";
 import DonationDonePage from "pages/donations/DonationDonePage";
-import DonationInProcessPage from "pages/donations/DonationInProcessPage";
 import ImpactPage from "pages/users/ImpactPage";
 import MainLayout from "layouts/MainLayout";
 import FundPage from "pages/promoters/FundPage";
@@ -44,14 +43,6 @@ function RoutesComponent(): JSX.Element {
         <Suspense fallback={<div />}>
           <CurrentUserProvider>
             <DonationDonePage />
-          </CurrentUserProvider>
-        </Suspense>
-      </Route>
-
-      <Route path="/donation-in-process" exact>
-        <Suspense fallback={<div />}>
-          <CurrentUserProvider>
-            <DonationInProcessPage />
           </CurrentUserProvider>
         </Suspense>
       </Route>
